+++ mev_inspect/events/__init__.py
@@ -1,0 +1,17 @@
+UNISWAP_V2 = "0xd78ad95fa46c994b6551d0da85fc275fe613ce37657fb8d5e3d130840159d822"
+UNISWAP_V3 = "0xc42079f94a6350d7e6235f29174924f928cc2ac818eb64fed8004e115fbcca67"
+AAVE = "0xe413a321e8681d831f4dbccbca790d2952b56f977908e45be37335533e005286"
+SYNAPSE = "0xc6c1e0630dbe9130cc068028486c0d118ddcea348550819defd5cb8c257f8a38"
+
+
+ALL_SWAPS = [
+    UNISWAP_V2,
+    UNISWAP_V3,
+    SYNAPSE,
+]
+
+ALL_LIQUIDATIONS = [
+    AAVE,
+]
+
+ALL_EVENTS = ALL_SWAPS + ALL_LIQUIDATIONS
